requirements.txt
Flask
pymongo
flask-cors
gunicorn
python-dotenv
